requirements.txt
asyncpraw
python-dotenv
langsmith
langchain
langchain_openai
langgraph
langgraph_supervisor
langchain_community
tavily-python
wikipedia-api
arxiv
youtube-transcript-api
google-api-python-client
wikipedia 
wikipedia-api 
nest_asyncio
asyncpraw
fastapi 
uvicorn 
httpx 
rich

